IOS-2446 Remove refactoring
